refactor: :recycle: convert plugin-window-unhandled-rejection to typescript and es modules
