+++ packages/plugin-window-unhandled-rejection/rollup.config.npm.mjs
@@ -1,0 +1,25 @@
+import createRollupConfig from '../../.rollup/index.mjs'
+
+export default createRollupConfig({
+    input: 'src/onerror.ts',
+    output: [
+        {
+            dir: `dist`,
+            entryFileNames: '[name].js',
+            format: 'cjs',
+            preserveModules: true,
+            generatedCode: {
+              preset: 'es2015',
+            }
+        },
+        {
+            dir: `dist`,
+            entryFileNames: '[name].mjs',
+            format: 'esm',
+            preserveModules: true,
+            generatedCode: {
+                preset: 'es2015',
+            }
+        }
+    ]
+})
