+++ packages/plugin-window-unhandled-rejection/src/unhandled-rejection.ts
@@ -1,21 +1,28 @@
-const map = require('@bugsnag/core/lib/es-utils/map')
-const isError = require('@bugsnag/core/lib/iserror')
+import { Plugin, Stackframe } from '@bugsnag/core'
+import map from '@bugsnag/core/lib/es-utils/map'
+import isError from '@bugsnag/core/lib/iserror'
 
-let _listener
+type Listener = (evt: PromiseRejectionEvent) => void
+
+let _listener: Listener | null
+
 /*
  * Automatically notifies Bugsnag when window.onunhandledrejection is called
  */
-module.exports = (win = window) => {
-  const plugin = {
+export default (win = window): Plugin => {
+  const plugin: Plugin = {
     load: (client) => {
+      // @ts-expect-error _config is private API
       if (!client._config.autoDetectErrors || !client._config.enabledErrorTypes.unhandledRejections) return
-      const listener = evt => {
+      const listener = (evt: PromiseRejectionEvent) => {
         let error = evt.reason
         let isBluebird = false
 
         // accessing properties on evt.detail can throw errors (see #394)
         try {
+          // @ts-expect-error detail does not exist on type PromiseRejectionEvent
           if (evt.detail && evt.detail.reason) {
+            // @ts-expect-error detail does not exist on type PromiseRejectionEvent
             error = evt.detail.reason
             isBluebird = true
           }
@@ -28,6 +35,7 @@ module.exports = (win = window) => {
           severity: 'error',
           unhandled,
           severityReason: { type: 'unhandledPromiseRejection' }
+          // @ts-expect-error _logger is private API
         }, 'unhandledrejection handler', 1, client._logger)
 
         if (isBluebird) {
@@ -49,7 +57,9 @@ module.exports = (win = window) => {
       if ('addEventListener' in win) {
         win.addEventListener('unhandledrejection', listener)
       } else {
+        // @ts-expect-error onunhandledrejection does not exist on type never
         win.onunhandledrejection = (reason, promise) => {
+          // @ts-expect-error detail does not exist on type PromiseRejectionEvent
           listener({ detail: { reason, promise } })
         }
       }
@@ -57,12 +67,14 @@ module.exports = (win = window) => {
     }
   }
 
+  // @ts-expect-error cannot find name 'process'
   if (process.env.NODE_ENV !== 'production') {
     plugin.destroy = (win = window) => {
       if (_listener) {
         if ('addEventListener' in win) {
           win.removeEventListener('unhandledrejection', _listener)
         } else {
+          // @ts-expect-error onunhandledrejection does not exist on type never
           win.onunhandledrejection = null
         }
       }
@@ -91,7 +103,7 @@ module.exports = (win = window) => {
 //
 // Bluebird pads method names with spaces so trim that too…
 // https://github.com/petkaantonov/bluebird/blob/b7f21399816d02f979fe434585334ce901dcaf44/src/debuggability.js#L568-L571
-const fixBluebirdStacktrace = (error) => (frame) => {
+const fixBluebirdStacktrace = (error: PromiseRejectionEvent['reason']) => (frame: Stackframe) => {
   if (frame.file === error.toString()) return
   if (frame.method) {
     frame.method = frame.method.replace(/^\s+/, '')
