+++ packages/plugin-window-unhandled-rejection/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../../tsconfig.json",
+    "include": ["src/**/*.ts"],
+    "compilerOptions": {
+      "target": "ES2020"
+    }
+}
+
