+++ tsconfig.json
@@ -81,7 +81,6 @@
     "packages/plugin-restify",
     "packages/node",
     "packages/plugin-client-ip",
-    "packages/plugin-window-unhandled-rejection",
     "packages/plugin-strip-query-string",
     "packages/plugin-strip-project-root",
     "packages/plugin-interaction-breadcrumbs",
